requirements.txt
anyio==3.6.2
bcrypt==4.0.1
beanie==1.18.0
bson==0.5.10
click==8.1.3
competitions-scheduler==0.2.5.post2
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.95.1
fixture==1.5.11
h11==0.14.0
httptools==0.5.0
idna==3.4
lazy-model==0.0.5
motor==3.1.2
numpy==1.24.3
pandas==2.0.1
passlib==1.7.4
Pillow==9.5.0
pyasn1==0.5.0
pydantic==1.10.7
pymongo==4.3.3
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
toml==0.10.2
tournament-scheduler==0.1.1
typing_extensions==4.5.0
tzdata==2023.3
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.2
